Dynamically create grid of square boxes
Create a function in JavaScript to create a square container of squares
that takes the grid size as input.

Default the grid size to 16x16 squares.
Retain the pixel size of the container regardless of the grid size.

Add a button to be used to dynamically create a grid based on input from
user.